# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Freelance Artist Finance",
    "version": "0.1.0"
  },
  "paths": {
...
    "/commissions": {
      "get": {
        "tags": [
          "commissions"
        ],
        "summary": "List Commissions",
        "operationId": "list_commissions_commissions_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CommissionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommissionPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "commissions"
        ],
        "summary": "Create Commission",
        "operationId": "create_commission_commissions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommissionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommissionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/commissions/{commission_id}": {
      "patch": {
        "tags": [
          "commissions"
        ],
        "summary": "Update Commission",
        "operationId": "update_commission_commissions__commission_id__patch",
        "parameters": [
          {
            "name": "commission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Commission Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommissionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommissionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "commissions"
        ],
        "summary": "Delete Commission",
        "operationId": "delete_commission_commissions__commission_id__delete",
        "parameters": [
          {
            "name": "commission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Commission Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommissionCreate": {
        "properties": {
          "client": {
            "type": "string",
            "minLength": 1,
            "title": "Client"
          },
          "piece": {
            "type": "string",
            "minLength": 1,
            "title": "Piece"
          },
          "hours_spent": {
            "type": "number",
            "minimum": 0.0,
            "title": "Hours Spent",
            "default": 0
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "expected_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Date"
          },
          "status": {
            "$ref": "#/components/schemas/CommissionStatus",
            "default": "in_progress"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Id"
          }
        },
        "type": "object",
        "required": [
          "client",
          "piece"
        ],
        "title": "CommissionCreate",
        "description": "API boundary: accepts dollars, exposes integer cents to the domain."
      },
      "CommissionPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CommissionRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "CommissionPage",
        "description": "Envelope pagination for commissions (see TransactionPage)."
      },
      "CommissionRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "client": {
            "type": "string",
            "title": "Client"
          },
          "piece": {
            "type": "string",
            "title": "Piece"
          },
          "hours_spent": {
            "type": "number",
            "title": "Hours Spent"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "expected_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Date"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Id"
          },
          "income_autologged": {
            "type": "boolean",
            "title": "Income Autologged"
          },
          "effective_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Rate"
          }
        },
        "type": "object",
        "required": [
          "id",
          "client",
          "piece",
          "hours_spent",
          "amount",
          "expected_date",
          "status",
          "transaction_id",
          "income_autologged",
          "effective_rate"
        ],
        "title": "CommissionRead"
      },
      "CommissionStatus": {
        "type": "string",
        "enum": [
          "agreed",
          "in_progress",
          "completed",
          "cancelled"
        ],
        "title": "CommissionStatus"
      },
...
      "CommissionUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/CommissionStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "CommissionUpdate",
        "description": "Terms (price, hours, expected date) lock when the commission is logged.\n\nOnly progress \u2014 the status \u2014 can be updated afterwards."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}